requirements.txt
amqp==2.5.2
billiard==3.6.1.0
cassandra-driver==3.20.1
celery==4.3.0
Cython==0.29.14
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
ez-setup==0.9
gunicorn==19.9.0
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
neo4j==1.7.6
neobolt==1.7.15
neotime==1.7.4
psycopg2==2.8.3
pyorient==1.5.5
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
